26-03|1:26

+++ src/graph.c
@@ -11,7 +11,7 @@ Graph *loadFile(char file[]){
         char in1[10], in2[30];
         fscanf(fd,"p edge %s %s",in1, in2);
         Graph *G = newGraph(atoi(in1));
-        int i, max = atoi(in2);
+        size_t i, max = atoi(in2);
         for(i=0;i<max;i++){
             fscanf(fd,"\ne %s %s",in1, in2);
             addEdge(G, atoi(in1), atoi(in2));
@@ -22,7 +22,7 @@ Graph *loadFile(char file[]){
     exit(1);
 }
 
-Graph *newGraph(short n_vertexs){
+Graph *newGraph(size_t n_vertexs){
     Graph *G = malloc(sizeof(Graph));
     G->v_count = n_vertexs;
     G->V = calloc(n_vertexs+1, sizeof(Vertex));
@@ -39,7 +39,7 @@ void addAdjinList(Adj_list *from_l,Vertex *to){
     from_l->last = n_adj;
 }
 
-void addEdge(Graph *G, short from, short to){
+void addEdge(Graph *G, size_t from, short to){
     addAdjinList(&G->V[from].adj, &G->V[to]);
     G->V[from].degree++;
     if(G->V[from].degree > G->high_degree)
